requirements.txt
Flask==2.0.1
jaydebeapi==1.2.3
mysql-connector-python==8.0.26
openai==0.27.0
